+++ harmonypy/__init__.py
@@ -1,4 +1,4 @@
 from .harmony import Harmony, run_harmony
 from .lisi import compute_lisi
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'